chore: Docs in Plugin Action Toolbar
