+++ app/client/src/PluginActionEditor/index.ts
@@ -13,3 +13,5 @@ export type {
 export { default as PluginActionNameEditor } from "./components/PluginActionNameEditor";
 
 export type { PluginActionEditorState } from "./store/pluginEditorReducer";
+
+export { DocsMenuItem } from "./components/PluginActionToolbar/components/DocsMenuItem";
