+++ app/client/src/ce/pages/Editor/AppPluginActionEditor/components/ToolbarMenu/index.ts
@@ -1,0 +1,1 @@
+export { ToolbarMenu } from "./ToolbarMenu";
